Accept flexible inputs when bools or numbers are expected.
When a bool is expected, we now accept the numbers 1 and 0, and the
case-insensitive strings "true", "false", "on", "off", "yes", "no",
"1", and "0".

When a number is expected, we now accept a string that parses to a
number.

New Value::to_flexible_{bool,num} functions.

For many configuration formats, it's convenient to accept the string
`"true"` is if it were the boolean `true`, or the string `"7"` as if it
were the number `7`.  This is part of making that possible.