Fix parsing of collections, short values.
This fixes a regression introduced in c730bdf in which an unterminated
collection string (e.g. "[") would cause the parser to loop infinitely.
It also fixes a regression in which the parser would discard the latter
part of a string if the earlier part parsed as a valid value while the
latter did not, e.g. " [discarded" in "okay [discarded".

Fixes #120.