Run cargo-fmt and add a CI action to check formatting
When attempting to make changes in the codebase, I noticed that the
automatic formatting on save meant that my editor was making large diffs
for small changes. This was because the code was not formatted
consistently. I ran cargo fmt to fix all this and added a CI action to
ensure that the code is always formatted correctly.